chore: upd naming
